backend/model/TextExtractor/requirements.txt
Flask>=2.0.0
google-generativeai
python-dotenv
